app/requirements.txt
flask~=3.1.1
aiogram~=3.20.0.post0
notion-client
pybit~=5.11.0
python-dotenv==0.19.0
waitress==2.1.2